Add Firebase support for leaderboard
I created a firebase, and populated it with random data. The leaderboard
page pulls from there.
Signing in doesn't actually add you to the database of users, but that
can easily be implemented.
This modifies gradle settings very significantly, so if gradle sync
doesn't work, clean the project under "Build -> Clean Project"