Add support to filter list of users from db by an array of userIds
This new change grabs a list of users' friends, and filters the raw list
of users from the db according to the user's friends. This effectively
results in only the user's friends being shown.
Unfortunately, since I have no friends (who use this app), I'm putting
in a sample array at the moment.